refactor: move EmailInput to the components directory
