+++ src/app/[lang]/signup/signUpForm.tsx
@@ -12,7 +12,7 @@ import {
 } from "@/util/input-validation";
 import { DictionaryReturnTypes } from "../dictionaries";
 import { LoginInput } from "./loginInput";
-import { EmailInput } from "../login/emailInput";
+import { EmailInput } from "@/components/input/EmailInput";
 
 export default function SignUpForm({
   lang,
